+++ lib/action_view/component/test_helpers.rb
@@ -3,8 +3,12 @@
 module ActionView
   module Component
     module TestHelpers
-      def render_component(component, &block)
-        Nokogiri::HTML(component.render_in(ApplicationController.new.view_context, &block))
+      def render_component(component, **args, &block)
+        if component.respond_to?(:render_in)
+          Nokogiri::HTML(component.render_in(ApplicationController.new.view_context, &block))
+        else
+          Nokogiri::HTML(component.new(args[:locals]).render_in(ApplicationController.new.view_context, &block))
+        end
       end
     end
   end