+++ test/app/controllers/application_controller.rb
@@ -3,4 +3,7 @@
 class ApplicationController < ActionController::Base
   def index
   end
+
+  def deprecated
+  end
 end