+++ test/config/routes.rb
@@ -2,4 +2,5 @@
 
 Dummy::Application.routes.draw do
   root to: "application#index"
+  get :deprecated, to: "application#deprecated"
 end