+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+*   Components are now rendered with `render MyComponent, locals: { foo: :bar }` syntax.
+
+    *Joel Hawksley*
+
 *   Components now inherit from ActionView::Component::base
 
     *Joel Hawksley*